rename loops.py to gen_loops.py
