+++ mmedit/engine/runner/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .loops import GenTestLoop, GenValLoop
+from .gen_loops import GenTestLoop, GenValLoop
 from .multi_loops import MultiTestLoop, MultiValLoop
 
 __all__ = ['MultiValLoop', 'MultiTestLoop', 'GenTestLoop', 'GenValLoop']
